requirements.txt
fastapi
uvicorn
ollama
pydantic-ai
pgvector
psycopg2-binary
pdf2image
pytesseract
numpy
pillow
torch
transformers
langchain
openai
python-multipart